⚡️ 优化（apple）：苹果认证  1、设置开机自动进入苹果认证发送探测包  2、修改传入参数len的类型为uint16_t  3、删除不必要的打印信息 结束

+++ applications/HL_A6902/utility/apple_auth/hl_util_apple.h
@@ -43,7 +43,7 @@ typedef enum _hl_apple_action_e_
 typedef struct _apple_func_handle_
 {
     /// 延时函数
-    void (*delay_usec)(uint16_t usec);
+    void (*delay_usec)(uint32_t usec);
     /// USB读数据函数
     int (*usb_read)(uint8_t* read_data_addr, uint16_t read_data_len, uint16_t timeout);
     /// USB写数据函数
